180Connect-bashir/backend/requirements.txt
fastapi[standard]
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
google-auth
ollama
requests
pandas
uvicorn